// upvertice-ui/src/app/services/models/page-response-providership-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ProvidershipResponse } from '../models/providership-response';
export interface PageResponseProvidershipResponse {
  content?: Array<ProvidershipResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}