// upvertice-ui/src/app/services/models/providership-multipart-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Wrapper for multipart providership creation/update request
 */
export interface ProvidershipMultipartRequest {

  /**
   * Proof images (optional)
   */
  images?: Array<Blob>;

  /**
   * Providership data as JSON string
   */
  request?: string;
}